+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
   buildToolsVersion "29.0.2"
 
   defaultConfig {
-    applicationId "com.doug.domain.challenge"
+    applicationId "com.doug.showcase"
     minSdkVersion 21
     targetSdkVersion 29
     versionCode 1
