+++ app/src/main/AndroidManifest.xml
@@ -1,17 +1,17 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.doug.domain.challenge">
+    package="com.doug.showcase">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".DomainChallengeApp"
+        android:name="com.doug.showcase.ShowcaseApp"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".ui.MainActivity">
+        <activity android:name="com.doug.showcase.ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <action android:name="android.intent.action.VIEW" />
