--- app/src/main/java/com/doug/domain/challenge/network/model/SearchRequest.kt
@@ -1,10 +1,0 @@
-package com.doug.domain.challenge.network.model
-
-import com.squareup.moshi.Json
-
-data class SearchRequest(
-    @Json(name = "dwelling_types")
-    val dwellingTypes: List<String>,
-    @Json(name = "search_mode")
-    val searchMode: String
-)
