--- app/src/main/java/com/doug/domain/challenge/repository/domain/DwellingType.kt
@@ -1,5 +1,0 @@
-package com.doug.domain.challenge.repository.domain
-
-enum class DwellingType {
-    APARTMENT, UNIT, FLAT
-}
