--- app/src/main/java/com/doug/domain/challenge/repository/domain/SearchType.kt
@@ -1,5 +1,0 @@
-package com.doug.domain.challenge.repository.domain
-
-enum class SearchType {
-    BUY, RENT
-}
