--- app/src/main/java/com/doug/domain/challenge/ui/MainActivity.kt
@@ -1,25 +1,0 @@
-package com.doug.domain.challenge.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.setupWithNavController
-import com.doug.domain.challenge.R
-import kotlinx.android.synthetic.main.activity_main.*
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        // Find the fragment navigation host and setup the app's navigation through
-        // the navigation graph (@navigation/navigation_graph)
-        val navController = findNavController(R.id.fragmentNavHost)
-        val configuration = AppBarConfiguration(navController.graph)
-        toolbar.setupWithNavController(navController, configuration)
-        setSupportActionBar(toolbar)
-    }
-
-}
