+++ app/src/main/java/com/doug/showcase/network/model/SearchRequest.kt
@@ -1,0 +1,10 @@
+package com.doug.showcase.network.model
+
+import com.squareup.moshi.Json
+
+data class SearchRequest(
+    @Json(name = "dwelling_types")
+    val dwellingTypes: List<String>,
+    @Json(name = "search_mode")
+    val searchMode: String
+)
