+++ app/src/main/java/com/doug/showcase/repository/domain/DwellingType.kt
@@ -1,0 +1,5 @@
+package com.doug.showcase.repository.domain
+
+enum class DwellingType {
+    APARTMENT, UNIT, FLAT
+}
