+++ app/src/main/java/com/doug/showcase/repository/domain/SearchType.kt
@@ -1,0 +1,5 @@
+package com.doug.showcase.repository.domain
+
+enum class SearchType {
+    BUY, RENT
+}
