+++ app/src/main/res/navigation/navigation_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/mainFragment"
-        android:name="com.doug.domain.challenge.ui.properties.PropertiesFragment"
+        android:name="com.doug.showcase.ui.properties.PropertiesFragment"
         android:label="@string/properties_title"
         tools:layout="@layout/fragment_properties" />
