+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name='DomainChallenge'
+rootProject.name='Showcase'
 include ':app'
